refactor: 페키지 구조 변경 (#337)
